requirements.txt
streamlit==1.26.0
pandas==2.0.3
plotly==5.17.0
requests==2.31.0
tqdm==4.66.1
google-generativeai==0.3.0
python-dotenv==1.0.0